### 2019-07-26 #63 @whedon
[PRE REVIEW]: FAIR_Bioinfo: a turnkey training course and protocol for reproducible computational biology
**Submitting author:** @thomasdenecker (<a href="http://orcid.org/0000-0003-1421-7641">Thomas DENECKER</a>)
**Repository:** <a href="https://github.com/thomasdenecker/FAIR_Bioinfo" target ="_blank">https://github.com/thomasdenecker/FAIR_Bioinfo</a>
**Version:** v1.2.0
