### 2019-08-12 #63 @labarba
@k8hertweck — Thank you for agreeing to review this submission for JOSE. We will want to find a second reviewer before starting the review (on a separate GitHub issue). But feel free to start looking at the submission, and browse our [Reviewer Guidelines](http://jose.theoj.org/about#reviewer_guidelines). More soon!
