### 2019-08-12 #68 @labarba
👋  @k8hertweck, @lexnederbragt — Thank you for agreeing to review for JOSE! This is where the action happens. You both have a review checklist to work through. Feel free to post any questions here. If you open new issues on the submission repository, please link to this issue here, and we'll get an auto-cross-link from GitHub. Godspeed!

[Reviewer guidelines](http://jose.theoj.org/about#reviewer_guidelines)
