### 2019-09-10 thomasdenecker/FAIR_Bioinfo#3 @lexnederbragt
Loading result files from container in the shiny app
I followed all the steps in https://github.com/thomasdenecker/FAIR_Bioinfo/blob/master/paper.md#how-to-use on my MacBook pro (I ran commands without `sudo`, though).

Running `docker exec -it fair_bioinfo bash ./FAIR_app.sh` allowed me to open `http://localhost:8888/rstudio/p/4444/` in the browser and I got the shiny app there. However, under 'Import', when I click on 'Browse' I guess am supposed to select files that are generated by the `FAIR_script.sh` script. When I click on the 'Browse' button, I can select files from my local filesystem. I am not very familiar with docker, so how do I get access to the relevant files (Conditions, Count table, Parameters)?

By the way, I had to use Google Chrome as browser, on Safari, I get an 'empty' shiny window.
