### 2019-09-10 #68 @lexnederbragt
The GitBook and course clearly address a pressing need for training in the Life Sciences. The book is comprehensive, and treats many, if not most of the tools that are relevant. I applaud the authors for doing such a thorough job of documenting their workflow for FAIR bioinformatics.

But herein also lies a challenge: the GitBook feels almost like a 'brain dump' of the authors. There are long descriptions of tools and only brief explanations on how to use them. Individual commands, choice of parameters, or alternative choices of these parameters are not explained, even though the taregt audience description says 'No prior computational skills are required to complete the entire course'. There is no lesson plan that helps others to design course activities, hardly any exercises. Is each lesson taught in one day? Over several days (which is what I think would be needed)? The exercises that are there are basically asking the learner to find a lot out on their own, there is very little guidance. There are no worked examples, although one could say the entire book is one huge worked example. I found no description of how the lessons were taught, the GitBook is a book, more than a lesson plan.

There is a lot of ground being covered and tools are often taught very superficially. As an example, the git lesson is around 3/4 page. In comparison (maybe a bit unfair, but still), the Software Carpentry git lesson takes 3 hours at least, and than it does not even cover branching!

So in all, I think the GitBook is an OK reference for self study by people familiar with the bioinformatic tools used. But FAIR_Bioinfo as it stands is not a stand-alone lesson that other teachers can easily adopt to their needs.

I would like to finish my review with a list of suggestions what can be done to improve the module:
- check the order of teaching tools, e.g. the shell command should be known when git is taught
- explain much more of what each command means, alternatively, assuming preexisting knowledge of the shell and/or the bioinformatic tools used
- provide the slides in English
- take it much slower, use more time per tool/part of the workflow
- add timings, and instructor nots: how is someone going to teach this section?
- add much more exercises, some scaffolded, and sparingly totally open exercises ('now run this on your own')
- prioritize, so with suggestions what instructors could skip if they have not enough time to teach the entire module
- make the learning outcomes for each chapter explicit
- help learners to install the necessary software to run the bioinformatics steps
- add contributor guidelines so that instructors using the module can give you input

----
I will now comment on why I did not mark some of boxes from the checklists.

### General checks

- **License:** Does the repository contain a plain-text LICENSE file with the contents of a standard license? (OSI-approved for code, Creative Commons for content)

The license is describes as `CeCILL FREE SOFTWARE LICENSE AGREEMENT`. I am unsure whether this license is compliant.

### Documentation

- **A statement of need:** Do the authors clearly state the need for this module and who the target audience is?

Only in the paper, not in the GitBook.

- **Installation instructions:** Is there a clearly stated list of dependencies?

Not in the GitBook, nor in the paper, and I did not find a `requirements.txt` file. However, there is a docker container.

- **Usage:** Does the documentation explain how someone would adopt the module, and include examples of how to use it?

There are instructions on how to run the entire analysis pipeline in the paper. I had one issue with that, see https://github.com/thomasdenecker/FAIR_Bioinfo/issues/3. There is little to no explanation on how to adopt the module.

- **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the module 2) Report issues or problems with the module 3) Seek support

No.

- **Learning objectives:** Does the module make the learning objectives plainly clear? (We don't require explicitly written learning objectives; only that they be evident from content and design.)

One can distill the learning objectives, but I hesitate to say that they are made 'plainly clear'.

- **Pedagogy:** Does the module seem easy to follow? Does it observe guidance on cognitive load? (working memory limits of 7 +/- 2 chunks of information)

No. There is a lot of information that is presented with fairly limited explanations. Cognitive load will be high. Examples:
    - https://fair-bioinfo.gitbook.io/fair-bioinfo/fair-principles#how-to-use-the-fair-data-principles does not introduce the shell and bash, jumps straight into git (`ls` in https://fair-bioinfo.gitbook.io/fair-bioinfo/code_memory is not explained, etc)
    - https://fair-bioinfo.gitbook.io/fair-bioinfo/some-first-definitions#what-is-a-terminal mentions terminal commands, but does not _teaches_ them, teaching happens _after_ git in https://fair-bioinfo.gitbook.io/fair-bioinfo/its-not-magic
    - https://fair-bioinfo.gitbook.io/fair-bioinfo/code_memory shows the use of `cd` which is not taught until next session/chapter
    - https://fair-bioinfo.gitbook.io/fair-bioinfo/some-first-definitions#a-couple-of-ssh-keys is going to lead to confusion, high cognitive load, could wait until it is needed rather than here
    - https://fair-bioinfo.gitbook.io/fair-bioinfo/its-not-magic: `wget`, `tar` and `sed` commands are not explained; also '2. wget (EBI): no installation required because wget is usually present by default.' wget is not installed by default on the Mac.
    - https://fair-bioinfo.gitbook.io/fair-bioinfo/its-not-magic#using-a-loop If you have never used a for loop in bash, this  `for j in $(tail -n +2 ../../conditions.txt)` is quite a big leap!
    - https://fair-bioinfo.gitbook.io/fair-bioinfo/its-not-magic uses `sudo apt-get update
    sudo apt-get install tree` to install tree, but this will not work on all systems (e.g., Mac); Iti would also help to explain `sudo`
    - https://fair-bioinfo.gitbook.io/fair-bioinfo/ive-got-the-power#creating-the-snakefile-of-our-project has the commands for the pipeline, but these are not run or explained outside of snakemake. As an example, what actual command is run, and what it does, could be explained: `bowtie2 -x {params.index} -U {input.file} -S {output.sam} > {output.out}`

- **Instructional design:** Is the instructional design deliberate and apparent? For example, exploit worked-example effects; effective multi-media use; low extraneous cognitive load.

There are no worked example; there are useful illustrative images, but not multi-media; at times, extraneous cognitive load is high (see also elsewhere). Can the slides be translated to English?

### JOSE paper

- Does it describe how it has been used in the classroom or other settings, and how someone might adopt it?

Maybe yes to the first, no to the second

- Could someone else teach with this module, given the right expertise?

Yes, but see comments elsewhere on 'book versus lesson plan'

- Does the paper tell the "story" of how the authors came to develop it, or what their expertise is?

Yes to the first (briefly), no to the second.

- **References:** Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?

I could find a link to a paper in the GitBook, but that did not have a doi; no doi's for datasets or software.

---------

Some minor issues:

## PDF
- Reproductible --> Reproducible
- the figure is misplaced, it appears between the heading of table 1 and table 1 itself

## https://github.com/thomasdenecker/FAIR_Bioinfo

- I ran the  workflow successfully without having to use `sudo`. Is running as root really necessary?

## GitBook
- There is quite some remaining French language, some of it should be translated
- https://fair-bioinfo.gitbook.io/fair-bioinfo/fair-principles#how-to-use-the-fair-data-principles "Thus, we do not follow the following principles" --> it looks like you _are_ following these principles?
- same page: "Tools used = references in their field" I am not sure what that means?
- https://fair-bioinfo.gitbook.io/fair-bioinfo/data-presentation --> could study system be introduced more? It is not even mentioned this is a green alga, and what the goal of the experiment was
- same page: fastq q-scores are not explained
- https://fair-bioinfo.gitbook.io/fair-bioinfo/some-first-definitions it does not become clear that the expectation is the learner installs a text editor
- https://fair-bioinfo.gitbook.io/fair-bioinfo/code_memory#what-when-why-git-and-github calls git a 'version manager' which I would not use (elsewhere it is called 'version control tool', which is OK)
- https://fair-bioinfo.gitbook.io/fair-bioinfo/code_memory#using-git repeatedly has `git statut`
- same page: 'Let's take the example of the family photo at the Sunday meal to illustrate the rest' I found that confusing and not an analogy that helps with understanding
- same page, the first `$ git diff` has a sentence that is not from its output: `The file will have its original line endings in your working directory.`
- https://fair-bioinfo.gitbook.io/fair-bioinfo/code_memory#how-to-create-a-new-branch-git-branch-less-than-newbranchname-greater-than `git branch` is not a git command I know, and as far as I know should be `git checkout -b <newBranchName>`
- same page, just below: `branche` --> `branch`
- https://fair-bioinfo.gitbook.io/fair-bioinfo/code_memory#how-to-switch-changes-from-one-branch-to-the-main-branch-git-merge-less-than-branchname-greater-than 'How to switch changes from one branch to the main branch?' --> I would not use 'switch' here
- https://fair-bioinfo.gitbook.io/fair-bioinfo/lover#table-exploration-with-dplyr is a weird stand-alone demo that learners probably think they will have to execute, but I think they did no load the data
- https://fair-bioinfo.gitbook.io/fair-bioinfo/notebooks#anaconda 'As regards reproducibility, doubts may persist.' --> I do not understand the meaning of this remark.
