### 2019-09-25 #68 @labarba
Thanks for the thorough review, @lexnederbragt 🙏

@thomasdenecker — Let us know your timeline for working on improvements based on the review.

@k8hertweck — I don't see any items in your checklist ticked off yet. When might you be able to start your review?
