### 2019-09-30 #68 @thomasdenecker
Hello
Thank you for the review.
I'm sorry to answer so late, but the teachings are taking up all my time. With @clairetn, we will try to answer your comments as soon as possible and resubmit a new proposal.
Thank you for your understanding,
