### 2019-10-20 #68 @labarba
👋  @thomasdenecker — Could we have a status update from you on your revision? When do you think you might be ready for the reviewers to take a second look? Let me know if you want me to set up a reminder here for you.
