### 2019-12-17 #68 @clairetn
I'm sorry to drag this long to make the suggested changes. I am planning a version for the new year, I hope that this deadline will be appropriate.
