### 2020-01-15 4038e53
Modifications after JOSE review
