### 2020-01-22 thomasdenecker/FAIR_Bioinfo#3 @thomasdenecker
In order to be able to use the file selector locally, you need a volume shared in Docker and the computer. To do this, add `-v` to docker command : `-v your_local_address:/srv/shiny-server`. For more details, the official documentation is here https://docs.docker.com/storage/volumes/ and in our Gitbook is here: https://app.gitbook.com/@fair-bioinfo/s/fair-bioinfo/a-trip-to-the-sea#image-to-container-how-to-launch-an-image-greater-than-docker-run
