### 2020-02-04 #68 @thomasdenecker
Hi everybody!

We come back to you with corrections for our submission. You will find on Github the updated paper and on gitbook, our educational content corrected according to the majority of the reviewer's requests. You will also find a point-by-point response letter attached concerning these requests.

We remain at your disposal,
Respectfully,

Thomas Denecker & Claire Toffano-Nioche (@clairetn)

[Letter_Denecker_Toffano-Nioche.pdf](https://github.com/openjournals/jose-reviews/files/4154699/Letter_Denecker_Toffano-Nioche.pdf)
