### 2020-04-09 #68 @thomasdenecker
Hi everyone,

We hope the confinement is not too difficult and that you and your family are healthy.
We take the liberty of contacting you to find out if you would like us to make changes to be published in JOSE.

We thank you for your feedback and remain at your disposal.

Thomas and Claire
