### 2020-04-30 #68 @lexnederbragt
My review of the resubmission:

I thank the authors for addressing many of my comments.
The book has now more of a workshop/course manual feeling,
with a lesson plan and timings (albeit coarse).
I'd be more inclined to use it if I were to teach these subjects.

I still feel some of my original comments are not fully addressed,
but I feel many of these can be resolved
by explicitly mentioning this for the reader.
For example, individual commands, choice of parameters,
or alternative choices of these parameters are still not explained.
The chosen workflow for RNAseq is given as is without it being taught.
It would help if the authors add that this explanation is not a part
of what the workshop intends to teach.

The given exercises are not very sophisticated.
Most lessons only have one exercise at the very end, asking the learner
to repeat stuff or add another thing.
This helps, but a lot of training is needed to become fluent with the tools
being taught.

Previous problems with the order of tools taught seem to be resolved.
Contributor guidelines still seem to be missing.
I recommend the authors to ask someone to proofread the English,
but this is not a showstopper.

I now also was able to run the whole pipeline using the commands
from the paper and look at the results on the shiny page.

## The Paper

Learning outcomes are missing the verbs: after this lesson,
learners 'are able to', 'know', etc.

## Gitbook https://fair-bioinfo.gitbook.io/fair-bioinfo/

* the actual commands for fastqc and samtools don't appear in the chapter
  "Play with analysis tools" but first in "I've got the power !"
  (it would be good to mention that in the former).
  However, in "I've got the power !", the actual commands are hidden 'inside'
  the snakefile. It would help the learner a lot if you *first* showed the
  actual shell command and what is does, *then* develop it in the snakefile,
  then explain the relation between the snakefile 'language'
  and the final command.
  Otherwise, it will be difficult for learners to develop snakefiles
  based on their own shell-commands/workflows
* the LoveR and Jupyter/Rmarkdown chapters feel very 'stand-alone',
  it does not become clear why they are there.
  It would help if they could be motivated by incorporating parts
  of the analysis to demonstrate their use in each chapter

----
I will now again comment on why I did not mark some of boxes from the checklists.

### General checks

- **License:** The `CeCILL FREE SOFTWARE LICENSE AGREEMENT` seems to be a software license, I recommend changing to a Creative Common license as advised by the journal.

### Documentation

- **Usage:** Does the documentation explain how someone would adopt the module, and include examples of how to use it?

There is still is little to no explanation on how to adopt the module.

- **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the module 2) Report issues or problems with the module 3) Seek support

No.

- **Pedagogy:** Does the module seem easy to follow? Does it observe guidance on cognitive load? (working memory limits of 7 +/- 2 chunks of information)

The module has improved on how easy it is to follow.
I still fear cognitive load will be high.

- **Instructional design:** Is the instructional design deliberate and apparent? For example, exploit worked-example effects; effective multi-media use; low extraneous cognitive load.

There are no worked examples; there are useful illustrative images, but not multi-media; at times, extraneous cognitive load is probably still high.

### JOSE paper

- Does it describe how it has been used in the classroom or other settings, and how someone might adopt it?

Maybe yes to the first, no to the second

- **References:** Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?

No doi's for datasets or software.

---------

**Some minor comments**

## GitBook

### Gain in bioinformatic analysis reproducibility as a goal

* change 2019 to 2020

### Our solution overview

The text
"If you are reading this, you are trying to answer the following question:
What are the solutions to be more reproducible in Bioinformatics?"

is repeated in the next sentence:

"If you are reading this, you are trying to answer the following question:
What are the solutions to be more reproducible in Bioinformatics?"

### Schedule
* Projet -> Project

### It's not magic

"Who's following?" Is not a good description of the `whoami` command...

"For the most courageous among you, it is quite possible to use the text editors available in the terminal such as VI, VIM, nano,... But we do not recommend it for beginners."

Please remove nano from that list as you are using it right after
Please add how to save the script generated in nano and exit nano

"As input data, we look for "RNAseq" data. They are composed of some files, each of them often counting more than a Go" -> I don't understand this sentence

### How to download web content with command line?
I suspect that this command `PATH /home/sratoolkit.2.9.2-ubuntu64/bin:$PATH`
should be `export PATH=/home/sratoolkit.2.9.2-ubuntu64/bin:$PATH`
(as you are talking about the bash shell)

### Play with analysis tools

Figure 'Reduction of a sam file by converting it into bam format' -> what is 'Mo'?
