### 2020-06-12 #68 @thomasdenecker
Dear @labarba , Dear all,

First of all, we would like to thank Lex Nederbragt for once again reviewing the current version of our work. We have tried to respond to the most important remarks and have made the majority of corrections. We appreciate the resulting improvement of our Gitbook. We would also like to thank Lorena for her help and assistance throughout the process and for following up on this review.

But today and for several reasons, we prefer to stop our collaboration and withdraw our request for publication in Jose. Thomas enters the traditionally busy period of the end of a thesis (with time constraints to finish the experiments, write the thesis manuscript and prepare the defense) and we have been also contacted by the French Institute of Bioinformatics to provide this training nationwide in September (https://www.france-bioinformatique.fr/fr/evenements/principes-fair-appliques-a-la-bioinformatique). As it requires a significant amount of preparation time, we will not be able to make any major changes to the publication we have submitted to you, above all that we don't still have any return of the 2nd reviewer.

We apologize for the time you spent there and thank you again for the rich process of improvement we shared with you.

Respectfully,

Thomas Denecker & Claire Toffano-Nioche
