### 2020-06-12 #68 @k8hertweck
Thank you for this update, and I want to apologize for my inability to provide comments. I've been struggling with health issues and it's clear I should not have obligated myself to this review when I've consistently not been able to provide the feedback necessary.
