### 2020-06-15 #68 @thomasdenecker
We are very sorry that you had health concerns. We understand that you were unable to review our proposal. We hope that your condition improves and that you will soon be able to resume your activities with peace of mind.

Respectfully,

Thomas Denecker & Claire Toffano-Nioche
