### 2020-06-17 #68 @lexnederbragt
Dear Thomas and Claire. Thanks for addressing the proposed changes (which I have not looked at yet). I hope that you will manage the coming busy period well and full understand that this prevents you from finalising this submission. I am mostly glad to have been of help.

And Kate, hope you will be well soon...

This was my first review for JOSE and I must say that it was a pleasant process, and feels like it was for a good cause (more open educational papers, and a refreshingly original and open review process).
