### 2020-11-24 #68 @labarba
hi @lexnederbragt — Phew! It has been one hell of a year. I come back to give some energy to JOSE after being almost crushed by the work as JupyterCon General Chair. Pivoting to an online event was truly a heavy lift.

Looking back through this review thread, I see that you have made a very thorough review. I would be comfortable accepting this submission on the basis of your recommendation, since Kate is unable to contribute. But I see that some of your items in the checklist are still unchecked. Are you satisfied with the revisions? Is there something else to do?
