### 2021-01-26 #68 @lexnederbragt
Given that the authors have indicated they no longer can put work into the paper, and in fact even write they want to "withdraw our request for publication in Jose", I've only looked at the issues I had outstanding. I feel that for me, the manuscript and gitbook can be published as is. But in the end, the journal/editor also has to decide that the work is acceptable to the journal, even with the remaining issues from your checklist and remaining reviewer comments.
