### 2021-05-23 #68 @labarba
hi @thomasdenecker 👋  — I see from following the links from your GitHub profile to your Twitter and to your website that you have graduated and are now officially a Doctor. Congratulations!

Would you like to come back to this JOSE submission and see if we can get this published? It got so close... I would be glad to help it through the finish line. Let me know!
