### 2021-05-27 #68 @thomasdenecker
hi @labarba,

Thank you very much ! A lot of work but a great achievement!

Regarding our submission, we could come back to it but could you please tell us what you are expecting specifically.

On our side, we are working in collaboration with the French Institute of Bioinformatics and we have done a session in September 2020 and soon in June 2021 :

- https://www.france-bioinformatique.fr/formation/fair-bioinfo-session-2/ ;
- https://ifb-elixirfr.github.io/IFB-FAIR-bioinfo-training/index.html#home

I thank you again for your help and encouragement throughout the process.
