### 2021-07-14 #68 @labarba
hi @thomasdenecker – I read through the paper, and have no further comments. The opinion of the reviewer (who gave extensive and thorough comments during this process) is that _"the manuscript and gitbook can be published as is"_. Therefore, at this time, the only things that remain are:

- [ ] make a tagged release on the repository, and report the version number here so I can update the metadata
- [ ] make a full deposit of the repository in Zenodo (or equivalent service), and report the archive DOI here, to enter it in the metadata
- [ ] manually edit the Zenodo archive title and author list as needed to match the JOSE paper
