### 2021-07-15 #68 @thomasdenecker
hi @labarba,

We thank you and we are delighted!

>  make a tagged release on the repository, and report the version number here so I can update the metadata

I just created it and the tag is: `JOSE publication` (2.0.0)

> make a full deposit of the repository in Zenodo (or equivalent service), and report the archive DOI here, to enter it in the metadata

The updated Zenodo DOI is : http://doi.org/10.5281/zenodo.5105349

> manually edit the Zenodo archive title and author list as needed to match the JOSE paper

Done

We thank you again and remain at your disposal.
