### 2021-07-25 #68 @labarba
@lexnederbragt 👋  — I am ready to accept this JOSE paper, on the basis of your extensive reviews and the authors' improvements. A few items of your checklist remain unchecked: could you have a final look? I think some could be easy to check off (e.g., License, References). Thank you!!
